# docs/swagger.yaml
basePath: /v1
definitions:
  dto.ConnectInstanceResponse:
    properties:
      alwaysOnline:
        type: boolean
      base64:
        type: string
      connected:
        type: boolean
      groupsIgnore:
        type: boolean
      id:
        type: string
      message:
        type: string
      msgCall:
        type: string
      proxyHost:
        type: string
      proxyPassword:
        type: string
      proxyPort:
        type: string
      proxyProtocol:
        type: string
      proxyUsername:
        type: string
      readMessages:
        type: boolean
      readStatus:
        type: boolean
      rejectCall:
        type: boolean
      remoteJID:
        type: string
      syncFullHistory:
        type: boolean
      syncRecentHistory:
        type: boolean
      webhook:
        $ref: '#/definitions/models.InstanceWebhook'
    type: object
...
  dto.CreateInstanceRequest:
    properties:
      alwaysOnline:
        type: boolean
      groupsIgnore:
        type: boolean
      id:
        type: string
      instanceName:
        type: string
      msgCall:
        type: string
      proxyHost:
        type: string
      proxyPassword:
        type: string
      proxyPort:
        type: string
      proxyProtocol:
        type: string
      proxyUsername:
        type: string
      readMessages:
        type: boolean
      readStatus:
        type: boolean
      rejectCall:
        type: boolean
      remoteJID:
        type: string
      syncFullHistory:
        type: boolean
      syncRecentHistory:
        type: boolean
      webhook:
        $ref: '#/definitions/models.InstanceWebhook'
    type: object
  dto.CreateInstanceResponse:
    properties:
      alwaysOnline:
        type: boolean
      groupsIgnore:
        type: boolean
      id:
        type: string
      msgCall:
        type: string
      proxyHost:
        type: string
      proxyPassword:
        type: string
      proxyPort:
        type: string
      proxyProtocol:
        type: string
      proxyUsername:
        type: string
      readMessages:
        type: boolean
      readStatus:
        type: boolean
      rejectCall:
        type: boolean
      remoteJID:
        type: string
      syncFullHistory:
        type: boolean
      syncRecentHistory:
        type: boolean
      webhook:
        $ref: '#/definitions/models.InstanceWebhook'
    type: object
  dto.DeleteInstanceResponse:
    properties:
      message:
        type: string
    type: object
...
  dto.ListInstancesResponse:
    properties:
      alwaysOnline:
        type: boolean
      groupsIgnore:
        type: boolean
      id:
        type: string
      instanceName:
        type: string
      msgCall:
        type: string
      ownerJid:
        type: string
      proxyHost:
        type: string
      proxyPassword:
        type: string
      proxyPort:
        type: string
      proxyProtocol:
        type: string
      proxyUsername:
        type: string
      readMessages:
        type: boolean
      readStatus:
        type: boolean
      rejectCall:
        type: boolean
      remoteJID:
        type: string
      syncFullHistory:
        type: boolean
      syncRecentHistory:
        type: boolean
      webhook:
        $ref: '#/definitions/models.InstanceWebhook'
    type: object
...
  dto.StatusInstanceResponse:
    properties:
      id:
        type: string
      instance:
        $ref: '#/definitions/dto.StatusInstanceResponseEvolutionCompatibility'
      state:
        type: string
    type: object
  dto.StatusInstanceResponseEvolutionCompatibility:
    properties:
      instanceName:
        type: string
      state:
        type: string
    type: object
  dto.UpdateInstanceRequest:
    properties:
      proxyHost:
        type: string
      proxyPassword:
        type: string
      proxyPort:
        type: string
      proxyProtocol:
        type: string
      proxyUsername:
        type: string
      webhook:
        properties:
          base64:
            type: boolean
          events:
            items:
              type: string
            type: array
          url:
            type: string
        type: object
    type: object
  dto.UpdateInstanceResponse:
    properties:
      alwaysOnline:
        type: boolean
      groupsIgnore:
        type: boolean
      id:
        type: string
      msgCall:
        type: string
      proxyHost:
        type: string
      proxyPassword:
        type: string
      proxyPort:
        type: string
      proxyProtocol:
        type: string
      proxyUsername:
        type: string
      readMessages:
        type: boolean
      readStatus:
        type: boolean
      rejectCall:
        type: boolean
      remoteJID:
        type: string
      syncFullHistory:
        type: boolean
      syncRecentHistory:
        type: boolean
      webhook:
        $ref: '#/definitions/models.InstanceWebhook'
    type: object
  models.InstanceWebhook:
    properties:
      base64:
        type: boolean
      byEvents:
        type: boolean
      enabled:
        type: boolean
      events:
        items:
          type: string
        type: array
      headers:
        additionalProperties:
          type: string
        type: object
      url:
        type: string
    type: object
  utils.HTTPErrorResponse:
    properties:
      error:
        type: string
      errorMessage:
        type: string
      message:
        type: string
    type: object
host: localhost:8080
info:
  contact:
    name: Verbeux AI
    url: https://github.com/verbeux-ai/whatsmiau
  description: WhatsMiau - WhatsApp API compatible with Evolution API. Provides instance
    management, messaging, and chat operations over WhatsApp Web.
  license:
    name: MIT
  title: WhatsMiau API
  version: 0.3.2
paths:
...
  /instance:
    get:
      description: Returns all instances, optionally filtered by name or ID
      parameters:
      - description: Filter by instance name
        in: query
        name: instanceName
        type: string
      - description: Filter by instance ID
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ListInstancesResponse'
            type: array
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List instances
      tags:
      - Instance
    post:
      consumes:
      - application/json
      description: Creates a new WhatsApp instance with the given name and optional
        configuration
      parameters:
      - description: Instance creation parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateInstanceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateInstanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new WhatsApp instance
      tags:
      - Instance
  /instance/{id}:
    delete:
      description: Disconnects and permanently removes the specified instance
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeleteInstanceResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete an instance
      tags:
      - Instance
  /instance/{id}/connect:
    post:
      description: Initiates connection for an instance. Returns a base64-encoded
        QR code PNG if not yet connected, or a connected status message.
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ConnectInstanceResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Connect an instance (get QR code)
      tags:
      - Instance
  /instance/{id}/logout:
    post:
      description: Disconnects the WhatsApp session for the given instance without
        deleting it
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeleteInstanceResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Logout an instance
      tags:
      - Instance
  /instance/{id}/status:
    get:
      description: Returns the current connection state of the specified instance
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StatusInstanceResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get instance connection state
      tags:
      - Instance
...
  /instance/connect/{id}:
    get:
      description: Initiates connection for an instance. Returns a base64-encoded
        QR code PNG if not yet connected, or a connected status message.
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.ConnectInstanceResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Connect an instance (get QR code)
      tags:
      - Instance
  /instance/connect/{id}/image:
    get:
      description: Returns the QR code as a raw PNG image buffer. Returns 204 No Content
        if already connected.
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - image/png
      responses:
        "200":
          description: QR code PNG image
          schema:
            type: file
        "204":
          description: Instance already connected
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get QR code as PNG image
      tags:
      - Instance
  /instance/connectionState/{id}:
    get:
      description: Returns the current connection state of the specified instance
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.StatusInstanceResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Get instance connection state
      tags:
      - Instance
  /instance/create:
    post:
      consumes:
      - application/json
      description: Creates a new WhatsApp instance with the given name and optional
        configuration
      parameters:
      - description: Instance creation parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.CreateInstanceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.CreateInstanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Create a new WhatsApp instance
      tags:
      - Instance
  /instance/delete/{id}:
    delete:
      description: Disconnects and permanently removes the specified instance
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeleteInstanceResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Delete an instance
      tags:
      - Instance
  /instance/fetchInstances:
    get:
      description: Returns all instances, optionally filtered by name or ID
      parameters:
      - description: Filter by instance name
        in: query
        name: instanceName
        type: string
      - description: Filter by instance ID
        in: query
        name: id
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.ListInstancesResponse'
            type: array
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: List instances
      tags:
      - Instance
  /instance/logout/{id}:
    delete:
      description: Disconnects the WhatsApp session for the given instance without
        deleting it
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.DeleteInstanceResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Logout an instance
      tags:
      - Instance
  /instance/update/{id}:
    put:
      consumes:
      - application/json
      description: Updates webhook and proxy settings for the given instance
      parameters:
      - description: Instance ID
        in: path
        name: id
        required: true
        type: string
      - description: Update parameters
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UpdateInstanceRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.UpdateInstanceResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "422":
          description: Unprocessable Entity
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/utils.HTTPErrorResponse'
      security:
      - ApiKeyAuth: []
      summary: Update an existing instance
      tags:
      - Instance
...
securityDefinitions:
  ApiKeyAuth:
    in: header
    name: apikey
    type: apiKey
swagger: "2.0"